much better i think now
